Adding types file and embedded CSS Handles
